Split fixtures in Shipping module
