+++ saleor/shipping/tests/fixtures/__init__.py
@@ -1,0 +1,4 @@
+from .shipping_method import *  # noqa: F401, F403
+from .shipping_method_data import *  # noqa: F401, F403
+from .shipping_method_translation import *  # noqa: F401, F403
+from .shipping_zone import *  # noqa: F401, F403
