+++ saleor/shipping/tests/fixtures/shipping_method_translation.py
@@ -1,0 +1,12 @@
+import pytest
+
+from ...models import ShippingMethodTranslation
+
+
+@pytest.fixture
+def shipping_method_translation_fr(shipping_method):
+    return ShippingMethodTranslation.objects.create(
+        language_code="fr",
+        shipping_method=shipping_method,
+        name="French shipping method name",
+    )
